refactor: 피드 페이지에서 사용하는 `Card` 컴포넌트 코로케이션 적용
