+++ src/pages/feed/components/BestIdeaCardListSection/BestIdeaCardListSection.tsx
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled';
 import { Spacer, Text } from 'concept-be-design-system';
 
-import BestIdeaCard from '../../../../components/Card/BestIdeaCard/BestIdeaCard';
+import BestIdeaCard from '../BestIdeaCard/BestIdeaCard';
 import { useBestIdeasQuery } from '../../hooks/queries/useBestIdeasQuery';
 import { useRef } from 'react';
 import { useFeedInfiniteFetch } from '../../hooks/useFeedInfiniteFetch';
