+++ src/pages/feed/components/NewIdeaCard/NewIdeaCardSkeleton.tsx
@@ -1,4 +1,4 @@
-import Skeleton from '../../Skeleton/Skeleton';
+import Skeleton from '../../../../components/Skeleton/Skeleton';
 
 const NewIdeaCardSkeleton = () => {
   return <Skeleton width="100%" height="300px" />;
